Break apart specs tests
